Restructuring SI_Toolkit (+ small corrections) Goal: easy switch between training in TF or Pytorch. Needed for: Delta GRU Done: Big part of TF functions rewritten/moved to serve also for Pytorch training while removing duplications. Still potential to remove duplicate code in Dataset.py. Additionaly small corrections to controllers: removing controller_nn_as_mpc_tf.py (duplication of controller_neural_imitator_tf.py) Correct import of lqr as auxiliary controller to mppi Deleting unused imports done semiautomatically, maybe unnecessarily much.
