+++ SI_Toolkit_ApplicationSpecificFiles_Template/run/Train_Network.py
@@ -1,0 +1,3 @@
+from SI_Toolkit.Training.Train import train_network
+
+train_network()
