--- SI_Toolkit_ApplicationSpecificFiles_Template/run/Train_TF.py
@@ -1,3 +1,0 @@
-from SI_Toolkit.TF.Train import train_tf
-
-train_tf()
