+++ src/SI_Toolkit/Functions/TF/TF2Numpy.py
@@ -13,7 +13,7 @@ from tensorflow.python.ops import gen_math_ops
 
 # These imports should be left
 import numpy as np
-from SI_Toolkit.TF.TF_Functions.Network import create_rnn_instance
+from SI_Toolkit.Functions.TF.Network import create_rnn_instance
 
 
 # Take care there is also other GRU implementation in TensorFlow
