+++ src/SI_Toolkit/Predictors/predictor_autoregressive_tf.py
@@ -36,11 +36,11 @@ Using predictor:
 # TODO: Make horizon updatable in runtime
 
 # "Command line" parameters
-from SI_Toolkit.TF.TF_Functions.Initialization import get_net, get_norm_info_for_net
-from SI_Toolkit.TF.TF_Functions.Normalising import get_normalization_function_tf, get_denormalization_function_tf, \
+from SI_Toolkit.Functions.General.Initialization import get_net, get_norm_info_for_net
+from SI_Toolkit.Functions.TF.Normalising import get_normalization_function_tf, get_denormalization_function_tf, \
     get_scaling_function_for_output_of_differential_network
-from SI_Toolkit.TF.TF_Functions.Network import _copy_internal_states_from_ref, _copy_internal_states_to_ref
-from SI_Toolkit.TF.TF_Functions.Compile import Compile
+from SI_Toolkit.Functions.TF.Network import _copy_internal_states_from_ref, _copy_internal_states_to_ref
+from SI_Toolkit.Functions.TF.Compile import Compile
 
 from SI_Toolkit_ASF.predictors_customization import STATE_VARIABLES, STATE_INDICES, \
     CONTROL_INPUTS
@@ -101,11 +101,11 @@ class predictor_autoregressive_tf(predictor):
         # Create a copy of the network suitable for inference (stateful and with sequence length one)
         self.net, self.net_info = \
             get_net(a, time_series_length=1,
-                    batch_size=self.batch_size, stateful=True)
+                    batch_size=self.batch_size, stateful=True, library='TF')
 
         net, _ = \
             get_net(a, time_series_length=1,
-                    batch_size=self.batch_size, stateful=True)
+                    batch_size=self.batch_size, stateful=True, library='TF')
 
         if np.any(['D_' in output_name for output_name in self.net_info.outputs]):
             self.differential_network = True
