--- src/SI_Toolkit/Pytorch/Train.py
@@ -1,406 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Jun 19 06:21:32 2020
-
-@author: Marcin
-"""
-
-import matplotlib.pyplot as plt
-
-import os
-
-import torch.nn as nn
-import torch.optim as optim
-import torch.utils.data.dataloader
-from torch.optim import lr_scheduler
-from tqdm import tqdm
-from torch.utils.tensorboard import SummaryWriter
-from torch.utils import data
-
-import timeit
-from warnings import warn as warning
-
-# "Command line" parameters
-from SI_Toolkit.load_parameters_for_training import args
-
-import time
-
-# Custom functions
-
-from SI_Toolkit.Pytorch.Pytorch_Functions.utilis import (
-    get_device,
-    set_seed,
-    get_net,
-    create_full_name,
-    get_norm_info_for_net,
-    create_log_file,
-    print_parameter_count,
-)
-
-from SI_Toolkit.Pytorch.Pytorch_Functions.Dataset import Dataset
-
-from SI_Toolkit.load_and_normalize import load_data, normalize_df, \
-    get_sampling_interval_from_datafile, get_paths_to_datafiles
-
-# region Import and print "command line" arguments
-print('')
-device = get_device()
-a = args()  # 'a' like arguments
-print(a.__dict__)
-print('')
-# endregion
-
-# Uncomment the @profile(precision=4) to get the report on memory usage after the training
-# Warning! It may affect performance. I would discourage you to use it for long training tasks
-# @profile(precision=4)
-def train_network(nni_parameters=None):
-    # region Start measuring time - to evaluate performance of the training function
-    start = timeit.default_timer()
-    # endregion
-
-    # region Set seeds to make experiment reproducible
-    set_seed(a)
-    # endregion
-
-    # region Make folder to keep trained models and their logs if not yet exist
-    try:
-        os.makedirs(a.path_to_models[:-1])
-    except FileExistsError:
-        pass
-    # endregion
-
-    net, net_info = get_net(a, device=device)
-
-    # Create a copy of the network suitable for inference (stateful and with sequence length one)
-    net_for_inference, net_for_inference_info = \
-        get_net(a, time_series_length=a.test_len,
-                batch_size=1, stateful=True, device=device)
-
-    # Create new full name for the pretrained net
-    create_full_name(net_info, a.path_to_models)
-    normalization_info = get_norm_info_for_net(net_info, files_for_normalization=a.training_files)
-    # а is an "argument"
-    # It must contain:
-    # path to models
-    # information about paths for:
-    #                              - training
-    #                              - validation
-    #                              - testing
-    create_log_file(net_info, a)
-
-    # region Load data and prepare datasets
-
-    if a.on_fly_data_generation:
-        ...
-        # TODO DatasetRandom should have normalization info too...
-        # TODO It should be possible only to provide here the dt as in norm info
-        # train_set = DatasetRandom(a, inputs_list=net_info.inputs, outputs_list=net_info.outputs, number_of_batches=1000)
-        # validation_set = DatasetRandom(a, inputs_list=net_info.inputs, outputs_list=net_info.outputs,
-        #                                number_of_batches=10)
-
-    else:
-
-        paths_to_datafiles_training = get_paths_to_datafiles(a.training_files)
-        paths_to_datafiles_validation = get_paths_to_datafiles(a.validation_files)
-
-        for path in paths_to_datafiles_training + paths_to_datafiles_validation:
-            try:
-                dt_sampling = get_sampling_interval_from_datafile(path)
-                if abs(net_info.sampling_interval - dt_sampling) > 1.0e-5:
-                    warning('A difference between network sampling interval and save interval of data file {} detected'
-                            .format(path))
-            except TypeError:
-                print('Sampling interval unknown.')
-
-        training_dfs = load_data(paths_to_datafiles_training)
-        validation_dfs = load_data(paths_to_datafiles_validation)
-
-        training_dfs_norm = normalize_df(training_dfs, normalization_info)
-        training_dataset = Dataset(training_dfs_norm, a, inputs=net_info.inputs, outputs=net_info.outputs)
-
-        validation_dfs_norm = normalize_df(validation_dfs, normalization_info)
-        validation_dataset = Dataset(validation_dfs_norm, a, inputs=net_info.inputs,
-                                     outputs=net_info.outputs)
-
-        # test_dfs is not deleted as we need it further for plotting
-        del training_dfs, validation_dfs, paths_to_datafiles_validation, paths_to_datafiles_training
-
-    # region In either case testing is done on a data collected offline
-    paths_to_datafiles_test = get_paths_to_datafiles(a.test_files)
-    test_dfs = load_data(paths_to_datafiles_test)
-    test_dfs_norm = normalize_df(test_dfs, normalization_info)
-    test_dataset = Dataset(test_dfs_norm, a, inputs=net_info.inputs, outputs=net_info.outputs)
-    # Check the sampling interval for test file
-    for path in paths_to_datafiles_test:
-        try:
-            dt_sampling = get_sampling_interval_from_datafile(path)
-            if abs(net_info.sampling_interval - dt_sampling) > 1.0e-5:
-                warning('A difference between network sampling interval and save interval of data file {} detected'
-                        .format(path))
-        except TypeError:
-            print('Sampling interval unknown')
-    # endregion
-
-    print('Number of samples in training set: {}'.format(training_dataset.number_of_samples))
-    print('The training sets sizes are: {}'.format(training_dataset.df_lengths))
-    print('Number of samples in validation set: {}'.format(validation_dataset.number_of_samples))
-    print('')
-
-    # endregion
-
-    # Create PyTorch dataloaders for train and dev set
-    training_generator = data.DataLoader(dataset=training_dataset, batch_size=a.batch_size, shuffle=True,
-                                      num_workers=a.num_workers)
-    validation_generator = data.DataLoader(dataset=validation_dataset, batch_size=512, shuffle=False, num_workers=a.num_workers)
-
-    # Print parameter count
-    print_parameter_count(net)  # Seems not to function well
-
-    # Select Optimizer
-    optimizer = optim.Adam(net.parameters(), amsgrad=True, lr=a.lr)
-
-    # TODO: Verify if scheduler is working. Try tweaking parameters of below scheduler and try cyclic lr scheduler
-    # scheduler = lr_scheduler.CyclicLR(optimizer, base_lr=lr, max_lr=0.1)
-    # scheduler = lr_scheduler.StepLR(optimizer, step_size=200, gamma=0.5)
-    scheduler = lr_scheduler.ReduceLROnPlateau(optimizer, 'min', patience=1, verbose=True)
-
-    # Select Loss Function
-    criterion = nn.MSELoss()  # Mean square error loss function
-
-    # Init Tensorboard
-    comment = f' batch_size={a.batch_size} lr={a.lr} exp_len={a.exp_len}'
-    tb = SummaryWriter(comment=comment, log_dir=net_info.path_to_net+'runs/')
-
-
-    ########################################################
-    # Training
-    ########################################################
-    print("Starting training...")
-    print('')
-    time.sleep(0.001)
-
-    # Create dictionary to store training history
-    dict_history = {}
-    dict_history['epoch'] = []
-    dict_history['time'] = []
-    dict_history['lr'] = []
-    dict_history['train_loss'] = []
-    dict_history['dev_loss'] = []
-    dict_history['dev_gain'] = []
-    dict_history['test_loss'] = []
-    dev_gain = 1
-
-    # The epoch_saved variable will indicate from which epoch is the last RNN model,
-    # which was good enough to be saved
-    epoch_saved = -1
-    for epoch in range(a.num_epochs):
-
-        ###########################################################################################################
-        # Training - Iterate batches
-        ###########################################################################################################
-        # Set RNN in training mode
-        net = net.train()
-        # Define variables accumulating training loss and counting training batchs
-        train_loss = 0
-        train_batches = 0
-
-        # Iterate training over available batches
-        # tqdm() is just a function which displays the progress bar
-        # Otherwise the line below is the same as "for batch, labels in train_generator:"
-        for batch, labels in tqdm(training_generator):  # Iterate through batches
-
-            # Reset the network (internal states of hidden layers and output history not the weights!)
-            net.reset()
-
-            # Further modifying the input and output form to fit RNN requirements
-            # If GPU available we send tensors to GPU (cuda)
-            if torch.cuda.is_available():
-                batch = batch.float().cuda().transpose(0, 1)
-                labels = labels.float().cuda()
-            else:
-                batch = batch.float().transpose(0, 1)
-                labels = labels.float()
-
-            # # Reset memory of gradients
-            # optimizer.zero_grad()
-
-            # Warm-up (open loop prediction) to settle the internal state of RNN hidden layers
-            net(rnn_input=batch[:a.wash_out_len, :, :])
-
-            # Reset memory of gradients
-            optimizer.zero_grad()
-
-            # Forward propagation - These are the results from which we calculate the update to RNN weights
-            # GRU Input size must be (exp_len, batch, input_size)
-            net(rnn_input=batch[a.wash_out_len:, :, :])
-            out = net.return_outputs_history()
-
-            # Get loss
-            loss = criterion(out[:, a.wash_out_len:, :],
-                             labels[:, a.wash_out_len:, :])
-
-            # Backward propagation
-            loss.backward()
-
-            # Gradient clipping - prevent gradient from exploding
-            torch.nn.utils.clip_grad_norm_(net.parameters(), 100)
-
-            # Update parameters
-            optimizer.step()
-            # scheduler.step()
-            # Update variables for loss calculation
-            batch_loss = loss.detach()
-            train_loss += batch_loss  # Accumulate loss
-            train_batches += 1  # Accumulate count so we can calculate mean later
-
-        ###########################################################################################################
-        # Validation - Iterate batches
-        ###########################################################################################################
-
-        # Set the network in evaluation mode
-        net = net.eval()
-
-        # Define variables accumulating evaluation loss and counting evaluation batches
-        dev_loss = 0
-        dev_batches = 0
-
-        for (batch, labels) in tqdm(validation_generator):
-
-            # Reset the network (internal states of hidden layers and output history not the weights!)
-            net.reset()
-
-            # Further modifying the input and output form to fit RNN requirements
-            # If GPU available we send tensors to GPU (cuda)
-            if torch.cuda.is_available():
-                batch = batch.float().cuda().transpose(0, 1)
-                labels = labels.float().cuda()
-            else:
-                batch = batch.float().transpose(0, 1)
-                labels = labels.float()
-
-            # Warm-up (open loop prediction) to settle the internal state of RNN hidden layers
-            net(rnn_input=batch)
-            out = net.return_outputs_history()
-
-
-            # Get loss
-            # For evaluation we always calculate loss over the whole maximal prediction period
-            # This allow us to compare RNN models from different epochs
-            loss = criterion(out[:, a.wash_out_len: a.exp_len],
-                             labels[:, a.wash_out_len: a.exp_len])
-
-            # Update variables for loss calculation
-            batch_loss = loss.detach()
-            dev_loss += batch_loss  # Accumulate loss
-            dev_batches += 1  # Accumulate count so we can calculate mean later
-
-        # Reset the network (internal states of hidden layers and output history not the weights!)
-        net.reset()
-        # Get current learning rate
-
-        for param_group in optimizer.param_groups:
-            lr_curr = param_group['lr']
-
-        scheduler.step(dev_loss)
-        '''
-        Add data for tensorboard
-        TODO : Add network graph and I/O to tensorboard
-        '''
-        # tb.add_graph(net)
-        tb.add_scalar('Train Loss', train_loss / train_batches, epoch)
-        tb.add_scalar('Dev Loss', dev_loss / dev_batches, epoch)
-
-        # Add the first sample of batch to tensorboard. Prediction is represented by Dotted line
-        # TODO: Concatenate such graphs. But they are not continous
-        # for i in range(labels.shape[2]):
-        #     time_label = np.arange(0, labels.shape[1], 1)
-        #     time_out = np.arange(0, out.shape[1], 1)
-        #     true_data = labels[1, :, i]
-        #     predicted_data = out[1, :, i]
-        #     fig_tb = plt.figure(5)
-        #     plt.plot(time_label, true_data.detach().cpu())
-        #     plt.plot(time_out, predicted_data.detach().cpu(), linestyle='dashed')
-        #     tb.add_figure(tag=str(a.outputs_list[i]), figure=fig_tb, global_step=epoch)
-
-        for name, param in net.named_parameters():
-            tb.add_histogram(name, param, epoch)
-            tb.add_histogram(f'{name}.grad', param.grad, epoch)
-        tb.close()
-
-        # Write the summary information about the training for the just completed epoch to a dictionary
-
-        dict_history['epoch'].append(epoch)
-        dict_history['lr'].append(lr_curr)
-        dict_history['train_loss'].append(
-            train_loss.detach().cpu().numpy() / train_batches / (a.exp_len - a.wash_out_len))
-        dict_history['dev_loss'].append(
-            dev_loss.detach().cpu().numpy() / dev_batches / (a.exp_len - a.wash_out_len))
-
-        # Get relative loss gain for network evaluation
-        if epoch >= 1:
-            dev_gain = (dict_history['dev_loss'][epoch - 1] - dict_history['dev_loss'][epoch]) / \
-                       dict_history['dev_loss'][epoch - 1]
-        dict_history['dev_gain'].append(dev_gain)
-
-        # Print the summary information about the training for the just completed epoch
-        print('\nEpoch: %3d of %3d | '
-              'LR: %1.5f | '
-              'Train-L: %6.4f | '
-              'Val-L: %6.4f | '
-              'Val-Gain: %3.2f |' % (dict_history['epoch'][epoch], a.num_epochs - 1,
-                                     dict_history['lr'][epoch],
-                                     dict_history['train_loss'][epoch],
-                                     dict_history['dev_loss'][epoch],
-                                     dict_history['dev_gain'][epoch] * 100))
-        print('')
-
-        # Save the best model with the lowest dev loss
-        # Always save the model from epoch 0
-        # TODO: this is a bug: you should only save the model from epoch 0 if there is no pretraind network
-        if epoch == 0:
-            min_dev_loss = dev_loss
-        # If current loss smaller equal than minimal till now achieved loss,
-        # save the current RNN model and save its loss as minimal ever achieved
-        if dev_loss <= min_dev_loss:
-            epoch_saved = epoch
-            min_dev_loss = dev_loss
-            torch.save(net.state_dict(), net_info.path_to_net + 'ckpt' + '.pt', _use_new_zipfile_serialization=False)
-            print('>>> saving best model from epoch {}'.format(epoch))
-            print('')
-        else:
-            print('>>> We keep model from epoch {}'.format(epoch_saved))
-            print('')
-
-        # Evaluate the performance of the current network
-        # by checking its predictions on a randomly generated CartPole experiment
-        # get_predictions(net, a, val_file)
-
-    # When finished the training print the final message
-    print("Training Completed...                                               ")
-    print(" ")
-
-    # region Calculate and print the total time it took to train the network
-
-    stop = timeit.default_timer()
-    total_time = stop - start
-
-    # Print the total time it took to run the function
-    print('Total time of training the network: ' + str(total_time))
-
-    # endregion
-
-
-if __name__ == '__main__':
-    import os.path
-    import time
-
-    # The following lines help to determine if the file on Google Colab was updated
-    file = os.path.realpath(__file__)
-    print("Training script last modified: %s" % time.ctime(os.path.getmtime(file)))
-
-    # Run the training function and measure time of execution
-    train_network()
-
-    # Use the call below instead of train_network() if you want to use NNI
-    # nni_parameters = nni.get_next_parameter()
-    # train_network(nni_parameters)
