+++ src/SI_Toolkit/Training/NNI/config.yml
@@ -17,6 +17,6 @@ tuner:
     #choice: maximize, minimize
     optimize_mode: minimize
 trial:
-  command: python -m Modeling.TF.Train --num_epochs=5
+  command: python -m SI_Toolkit.TF.Train --num_epochs=5
   codeDir: ../../../
   gpuNum: 0
